# swagger.yaml
openapi: 3.0.3
info:
  title: MEDAPP API
  description: MEDAPP API
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local server
...
paths:
...
  /api/v1/doctors/{id}:
    get:
      tags:
        - Doctors
      summary: Get doctor by ID
      operationId: getDoctorById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctor details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TResult"
                  - properties:
                      data:
                        $ref: "#/components/schemas/Doctor"
        "404":
          description: Doctor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResult"
    put:
      tags:
        - Doctors
      summary: Update doctor
      operationId: updateDoctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDoctorRequest"
      responses:
        "200":
          description: Doctor updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TResult"
                  - properties:
                      data:
                        $ref: "#/components/schemas/Doctor"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResult"
  /api/v1/doctors:
    get:
      tags:
        - Doctors
      summary: Get all doctors
      operationId: getAllDoctors
      responses:
        "200":
          description: List of doctors
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TResult"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Doctor"
    post:
      tags:
        - Doctors
      summary: Create a new doctor
      operationId: createDoctor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDoctorRequest"
      responses:
        "201":
          description: Doctor created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TResult"
                  - properties:
                      data:
                        $ref: "#/components/schemas/Doctor"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResult"
...
  /api/v1/appointments:
    get:
      tags:
        - Appointments
      summary: Get all appointments
      operationId: getAllAppointments
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      tags:
        - Appointments
      summary: Create a new appointment
      operationId: createAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentRequest"
      responses:
        "201":
          description: Appointment created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /api/v1/appointments/{id}:
    get:
      tags:
        - Appointments
      summary: Get appointment by ID
      operationId: getAppointmentById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    put:
      tags:
        - Appointments
      summary: Update appointment
      operationId: updateAppointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentRequest"
      responses:
        "200":
          description: Appointment updated successfully
    delete:
      tags:
        - Appointments
      summary: Delete appointment
      operationId: removeAppointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment deleted successfully
  /api/v1/appointments/patient/{patientId}:
    get:
      tags:
        - Appointments
      summary: Get appointments by patient ID
      operationId: getAppointmentByPatientId
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
  /api/v1/appointments/doctor/{doctorId}:
    get:
      tags:
        - Appointments
      summary: Get appointments by doctor ID
      operationId: getAppointmentByDoctorId
      parameters:
        - name: doctorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
  /api/v1/emergencyContact/{patientId}:
    get:
      tags:
        - EmergencyContact
      summary: Get emergency contacts by patient ID
      operationId: getByPatientIdController
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of emergency contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmergencyContact"
  /api/v1/emergencyContact:
    post:
      tags:
        - EmergencyContact
      summary: Create a new emergency contact
      operationId: createEmergencyContact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmergencyContactRequest"
      responses:
        "201":
          description: Emergency contact created successfully
  /api/v1/emergencyContact/{id}:
    put:
      tags:
        - EmergencyContact
      summary: Update emergency contact
      operationId: updateEmergencyContact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmergencyContactRequest"
      responses:
        "200":
          description: Emergency contact updated successfully
    delete:
      tags:
        - EmergencyContact
      summary: Delete emergency contact
      operationId: removeEmergencyContact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Emergency contact deleted successfully
...
  /api/v1/vitalSign/appointment/{id}:
    get:
      tags:
        - VitalSign
      summary: Get vital sign by appointment ID
      operationId: getByAppointmentId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vital sign details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VitalSign"
...
  /api/v1/medical-consultations:
    get:
      tags:
        - MedicalConsultation
      summary: Get all medical consultations
      operationId: getAllConsultations
      responses:
        "200":
          description: List of medical consultations
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TResult"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/MedicalConsultation"
    post:
      tags:
        - MedicalConsultation
      summary: Create a new medical consultation
      operationId: createConsultation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedicalConsultationRequest"
      responses:
        "201":
          description: Medical consultation created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TResult"
                  - properties:
                      data:
                        $ref: "#/components/schemas/MedicalConsultation"
  /api/v1/medical-consultations/{id}:
    get:
      tags:
        - MedicalConsultation
      summary: Get medical consultation by ID
      operationId: getConsultationById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medical consultation details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TResult"
                  - properties:
                      data:
                        $ref: "#/components/schemas/MedicalConsultation"
        "404":
          description: Medical consultation not found
    put:
      tags:
        - MedicalConsultation
      summary: Update medical consultation
      operationId: updateConsultation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMedicalConsultationRequest"
      responses:
        "200":
          description: Medical consultation updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TResult"
                  - properties:
                      data:
                        $ref: "#/components/schemas/MedicalConsultation"
    delete:
      tags:
        - MedicalConsultation
      summary: Delete medical consultation
      operationId: removeConsultation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medical consultation deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResult"
  /api/v1/medical-consultations/patient/{patientId}:
    get:
      tags:
        - MedicalConsultation
      summary: Get medical consultations by patient ID
      operationId: getConsultationsByPatientId
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of medical consultations
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TResult"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/MedicalConsultation"
  /api/v1/medical-consultations/appointment/{appointmentId}:
    get:
      tags:
        - MedicalConsultation
      summary: Get medical consultation by appointment ID
      operationId: getConsultationByAppointmentId
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medical consultation details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TResult"
                  - properties:
                      data:
                        $ref: "#/components/schemas/MedicalConsultation"
        "404":
          description: Medical consultation not found
...
components:
  schemas:
    TResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          nullable: true
...
    Doctor:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        lastName:
          type: string
        email:
          type: string
        licenseNumber:
          type: string
        specialty:
          type: string
        subSpecialty:
          type: string
        phone:
          type: string
        officePhone:
          type: string
    CreateDoctorRequest:
      type: object
      required:
        - email
        - password
        - name
        - lastName
        - licenseNumber
        - specialty
        - phone
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        name:
          type: string
        lastName:
          type: string
        licenseNumber:
          type: string
        specialty:
          type: string
        subSpecialty:
          type: string
        phone:
          type: string
        officePhone:
          type: string
    UpdateDoctorRequest:
      type: object
      properties:
        name:
          type: string
        lastName:
          type: string
        specialty:
          type: string
        subSpecialty:
          type: string
        phone:
          type: string
        officePhone:
          type: string
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date-time
        time:
          type: string
        reason:
          type: string
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        patientId:
          type: integer
        doctorId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAppointmentRequest:
      type: object
      required:
        - date
        - time
        - reason
        - status
        - patientId
        - doctorId
      properties:
        date:
          type: string
          format: date-time
        time:
          type: string
        reason:
          type: string
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        patientId:
          type: integer
        doctorId:
          type: integer
    UpdateAppointmentRequest:
      type: object
      properties:
        date:
          type: string
          format: date-time
        time:
          type: string
        reason:
          type: string
        status:
          $ref: "#/components/schemas/AppointmentStatus"
    AppointmentStatus:
      type: string
      enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED]
    EmergencyContact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        patientId:
          type: integer
    CreateEmergencyContactRequest:
      type: object
      required:
        - name
        - lastName
        - phone
        - email
        - address
        - patientId
      properties:
        name:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        patientId:
          type: integer
    UpdateEmergencyContactRequest:
      type: object
      properties:
        name:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
    VitalSign:
      type: object
      properties:
        id:
          type: integer
        temperature:
          type: number
        heartRate:
          type: integer
        bloodPressure:
          type: string
        oxygenSaturation:
          type: integer
        respiratoryRate:
          type: integer
        weight:
          type: number
        height:
          type: number
        appointmentId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    MedicalConsultation:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        doctorId:
          type: integer
        appointmentId:
          type: integer
        symptoms:
          type: string
        physicalExam:
          type: string
        diagnosis:
          type: string
        treatment:
          type: string
        recommendations:
          type: string
          nullable: true
        followUp:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateMedicalConsultationRequest:
      type: object
      required:
        - patientId
        - doctorId
        - appointmentId
        - symptoms
        - physicalExam
        - diagnosis
        - treatment
      properties:
        patientId:
          type: integer
        doctorId:
          type: integer
        appointmentId:
          type: integer
        symptoms:
          type: string
        physicalExam:
          type: string
        diagnosis:
          type: string
        treatment:
          type: string
        recommendations:
          type: string
          nullable: true
        followUp:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    UpdateMedicalConsultationRequest:
      type: object
      properties:
        symptoms:
          type: string
        physicalExam:
          type: string
        diagnosis:
          type: string
        treatment:
          type: string
        recommendations:
          type: string
          nullable: true
        followUp:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true